chore: change route path name of mypage from mypage/home to mypage
